fixing tests

+++ main/utils.py
@@ -357,7 +357,8 @@ def clean_data(data: str) -> str:
 
 def clear_search_cache():
     cache = caches["redis"]
-    search_keys = cache.keys("views.decorators.cache.cache_page.search.*")
-    cache.delete_many(search_keys)
-    search_keys = cache.keys("views.decorators.cache.cache_header.search.*")
-    cache.delete_many(search_keys)
+    if hasattr(cache, "keys"):
+        search_keys = cache.keys("views.decorators.cache.cache_page.search.*")
+        cache.delete_many(search_keys)
+        search_keys = cache.keys("views.decorators.cache.cache_header.search.*")
+        cache.delete_many(search_keys)
